+++ .idea/workspace.xml
@@ -3,10 +3,14 @@
   <component name="ChangeListManager">
     <list default="true" id="048a2705-6e2e-4c94-81f1-7e80bed125e6" name="Changes" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/App/App.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/App/App.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/App/components/Header/Header.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/App/components/Header/Header.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/App/components/SearchBox/SearchBox.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/App/components/SearchBox/SearchBox.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/App/components/SearchBox/travelersApi.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/App/components/SearchBox/travelersApi.ts" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/App/components/User/Auth.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/App/components/User/Auth.tsx" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/App/components/User/SignIn.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/App/components/User/SignIn.tsx" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/App/components/layout/MainNavigation.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/App/components/layout/MainNavigation.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/App/store/auth.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/App/store/createAuthSlice.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/App/store/store.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/App/store/store.ts" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -45,7 +49,7 @@
       <updated>1661333936747</updated>
       <workItem from="1661333938401" duration="9000" />
       <workItem from="1661333951824" duration="6512000" />
-      <workItem from="1661454065146" duration="3466000" />
+      <workItem from="1661454065146" duration="8040000" />
     </task>
     <servers />
   </component>
