+++ src/App/components/Header/Header.tsx
@@ -5,7 +5,7 @@ import { useSelector, useDispatch } from 'react-redux';
 import Image from 'App/components/Reuse/Image';
 import Button from 'App/components/Reuse/Button';
 import classes from 'App/components/Header/Header.module.scss';
-import { authActions } from '../../store/auth';
+import { authActions } from 'App/store/createAuthSlice';
 
 function Header() {
   const dispatch = useDispatch();
