+++ src/App/components/User/Auth.tsx
@@ -4,7 +4,7 @@ import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
 import Input from 'App/components/Reuse/Input';
 import classes from 'App/components/User/Auth.module.scss';
-import { authActions } from 'App/store/auth';
+import { authActions } from 'App/store/createAuthSlice';
 import Button from 'App/components/Reuse/Button';
 import { useNavigate } from 'react-router-dom';
 import { useCookies } from 'react-cookie';
