+++ src/App/components/User/SignIn.tsx
@@ -6,7 +6,7 @@ import classes from 'App/components/User/SignIn.module.scss';
 import Input from 'App/components/Reuse/Input';
 import Button from 'App/components/Reuse/Button';
 import { useNavigate } from 'react-router-dom';
-import { authActions } from 'App/store/auth';
+import { authActions } from 'App/store/createAuthSlice';
 import { useDispatch } from 'react-redux';
 import { useCookies } from 'react-cookie';
