+++ src/App/components/layout/MainNavigation.tsx
@@ -5,7 +5,7 @@ import Button from 'App/components/Reuse/Button';
 import React from 'react';
 import {useDispatch, useSelector} from 'react-redux';
 import Image from 'App/components/Reuse/Image';
-import {authActions} from 'App/store/auth';
+import {authActions} from 'App/store/createAuthSlice';
 import classes from 'App/components/layout/MainNavigation.module.scss';
 import logo from 'logo-1.png';
